mt-aaa-be-api/requirements.txt
celery==5.2.2
fastapi==0.65.2
future==0.18.2
gunicorn==20.0.4
h11==0.9.0
httptools==0.1.1
pydantic==1.6.2
pytz==2020.1
redis==3.5.3
uvicorn
uvloop==0.14.0
python-dotenv
pandas
numpy
celery
redis
scikit-learn
asyncpg==0.22.0
ormar==0.10.5
psycopg2-binary==2.8.6
requests